+++ src/build32.sh
@@ -8,5 +8,5 @@ echo ./gen mem.s
 ./gen mem.s
 fi
 
-echo gcc -m32 -g3 gc.c flow.c lisp.c cell.c mem32.c math.c -o lisp32
-gcc -m32 -g3 gc.c flow.c lisp.c cell.c mem32.c math.c -o lisp32
+echo gcc -m32 -g3 sym.c read.c gc.c flow.c lisp.c cell.c mem32.c math.c -o lisp32
+gcc -m32 -g3 sym.c read.c gc.c flow.c lisp.c cell.c mem32.c math.c -o lisp32
