+++ src/build64.sh
@@ -8,5 +8,5 @@ echo ./gen mem.s
 ./gen mem.s
 fi
 
-echo gcc -g3 gc.c flow.c lisp.c cell.c mem64.c math.c -o lisp64
-gcc -g3 gc.c flow.c lisp.c cell.c mem64.c math.c -o lisp64
+echo gcc -g3 sym.c read.c gc.c flow.c lisp.c cell.c mem64.c math.c -o lisp64
+gcc -g3 sym.c read.c gc.c flow.c lisp.c cell.c mem64.c math.c -o lisp64
