+++ src/lisp.c
@@ -56,354 +56,6 @@ any ApplyArgs, ApplyBody;
 //               sym.c
 ///////////////////////////////////////////////
 
-int firstByte(any s)
-{
-    int c;
-
-    if (isNil(s)) return 0;
-    c = (uword)s;
-    return c & 127;
-}
-
-int getByte1(int *i, uword *p, any *q)
-{
-    int c;
-
-    if (getCARType(*q) == TXT)
-    {
-        (*q)=(*q)->car;
-        *i = BITS, *p = (uword)(*q) , *q = NULL;
-    }
-    else if (getCARType(*q) == BIN_START)
-    {
-        (*q)=(*q)->car;
-        *i = BITS, *p = (uword)((*q)->car) , *q = ((*q)->cdr);
-    }
-    else
-    {
-        giveup("Cant getByte");
-    }
-
-    c = *p & 127, *p >>= 8, *i -= 8;
-
-    return c;
-}
-
-int getByte(int *i, uword *p, any *q)
-{
-    int c;
-
-    if (*i == 0)
-    {
-        if (!*q)
-        {
-            return 0;
-        }
-        else
-        {
-            *i = BITS,  *p = (uword)((*q)->car),  *q = (*q)->cdr;
-        }
-    }
-    c = *p & 127,  *p >>= 8;
-    if (*i >= 8)
-        *i -= 8;
-    else if (isNum(*q))
-    {
-        *p = (uword)*q >> 2,  *q = NULL;
-        c |= *p << *i;
-        *p >>= 8 - *i;
-        *i += BITS-9;
-    }
-    else
-    {
-        *p = (uword)tail(*q),  *q = val(*q);
-        c |= *p << *i;
-        *p >>= 8 - *i;
-        *i += BITS-8;
-    }
-    c &= 127;
-
-    return c;
-}
-
-any mkTxt(int c)
-{
-    return txt(c & 127);
-}
-
-any mkChar(int c)
-{
-   return consSym(NULL, c & 127);
-}
-
-void putByte1(int c, int *i, uword *p, any *q)
-{
-    *p = c & 127;
-    *i = 8;
-    *q = NULL;
-}
-
-void putByte(int c, int *i, uword *p, any *q, cell *cp)
-{
-    c = c & 127;
-    int d = 8;
-
-    if (*i != BITS)
-        *p |= (uword)c << *i;
-
-    if (*i + d  > BITS)
-    {
-        if (*q)
-        {
-            any x = consName(*p, Zero);
-            setCARType(x, BIN);
-            (*q)->cdr = x;
-            *q = x;
-        }
-        else
-        {
-            any x = consSym(NULL, 0);
-            setCARType(x, BIN_START);
-            Push(*cp, x);
-            any y = consName(*p, Zero);
-            setCARType(y, BIN);
-            (*cp).car->car = *q = y;
-
-        }
-        *p = c >> BITS - *i;
-        *i -= BITS;
-    }
-
-    *i += d;
-}
-
-any popSym(int i, uword n, any q, cell *cp)
-{
-    if (q)
-    {
-        //val(q) = i <= (BITS-2)? box(n) : consName(n, Zero);
-        q->cdr = consName(n, Nil);
-        return Pop(*cp);
-    }
-    return consSym(NULL,n);
-}
-
-int symBytes(any x)
-{
-    int cnt = 0;
-    uword w;
-
-    if (isNil(x))
-        return 0;
-
-    if (isTxt(x))
-    {
-        w = (uword)(x->car);
-        while (w)
-        {
-            ++cnt;
-            w >>= 8;
-        }
-    }
-
-    return cnt;
-}
-
-any isIntern(any nm, any tree[2])
-{
-    any x, y, z;
-    word n;
-
-    if (isTxt(nm))
-    {
-        for (x = tree[0];  x != Nil;)
-        {
-            if ((n = (word)(car(nm)) - (word)name(caar(x))) == 0)
-            {
-                return car(x);
-            }
-            x = n<0? cadr(x) : cddr(x);
-        }
-    }
-    else
-    {
-        for (x = tree[1];  x != Nil;)
-        {
-            y = nm->car;
-            z = x->car->car;
-            while ((n = (word)(car(y)) - (word)car(z)) == 0)
-            {
-                if (getCARType(y) != BIN) return car(x);
-                y=y->cdr;
-                z=z->cdr;
-            }
-
-            x = n<0? cadr(x) : cddr(x);
-        }
-    }
-
-
-
-    return NULL;
-}
-
-any internBin(any sym, any tree[2])
-{
-    any nm, x, y, z;
-    word n;
-
-    x = tree[1];
-
-    if (x == Nil)
-    {
-        tree[1] = consIntern(sym, Nil);
-        return tree[1];
-    }
-
-    for (;;)
-    {
-
-        y = sym->car;
-        z = x->car->car;
-        while ((n = (word)(car(y)) - (word)car(z)) == 0)
-        {
-            if (getCARType(y) != BIN) return sym;
-            y=y->cdr;
-            z=z->cdr;
-        }
-
-        if (Nil == cdr(x))
-        {
-            cdr(x) = n < 0 ? consIntern(consIntern(sym, Nil), Nil) : consIntern(Nil, consIntern(sym, Nil));
-            return sym;
-        }
-        if (n < 0)
-        {
-            if (Nil != cadr(x))
-            {
-                x = cadr(x);
-            }
-            else
-            {
-                cadr(x) = consIntern(sym, Nil);
-                return sym;
-            }
-        }
-        else
-        {
-            if (Nil != cddr(x))
-            {
-                x = cddr(x);
-            }
-            else
-            {
-                cddr(x) = consIntern(sym, Nil);
-                return sym;
-            }
-        }
-    }
-}
-
-any intern(any sym, any tree[2])
-{
-   any nm, x;
-   word n;
-
-
-   if (getCARType(sym) == BIN_START) return internBin(sym, tree);
-
-   nm = sym;
-
-   x = tree[0];
-   if (Nil == x)
-   {
-      tree[0] = consIntern(sym, Nil);
-      return tree[0];
-   }
-   for (;;)
-   {
-      if ((n = (word)(car(nm)) - (word)name(caar(x))) == 0)
-      {
-         return car(x);
-      }
-
-      if (Nil == cdr(x))
-      {
-         cdr(x) = n < 0 ? consIntern(consIntern(sym, Nil), Nil) : consIntern(Nil, consIntern(sym, Nil));
-         return sym;
-      }
-      if (n < 0)
-      {
-         if (Nil != cadr(x))
-         {
-            x = cadr(x);
-         }
-         else
-         {
-            cadr(x) = consIntern(sym, Nil);
-            return sym;
-         }
-      }
-      else
-      {
-         if (Nil != cddr(x))
-         {
-            x = cddr(x);
-         }
-         else
-         {
-            cddr(x) = consIntern(sym, Nil);
-            return sym;
-         }
-      }
-   }
-}
-
-/* Get symbol name */
-any name(any s)
-{
-   return s;
-}
-
-/* Make name */
-any mkSym(byte *s)
-{
-    int i;
-    uword w;
-    cell c1, *p;
-
-    putByte1(*s++, &i, &w, &p);
-    while (*s)
-    {
-        putByte(*s++, &i, &w, &p, &c1);
-    }
-    return popSym(i, w, p, &c1);
-}
-
-/* Make string */
-any mkStr(char *s)
-{
-   if (s && *s)
-   {
-      return mkSym((byte *)s);
-   }
-   else
-   {
-      return Nil;
-   }
-}
-
-// (==== ['sym ..]) -> NIL
-any doHide(any ex)
-{
-    // TODO - is this needed?
-   printf("%p\n", ex);
-
-   return Nil;
-}
-
-void makeError(any ex) {err(ex, NULL, "Not making");}
-
 
 ///////////////////////////////////////////////
 //               sym.c - END
@@ -414,468 +66,7 @@ void makeError(any ex) {err(ex, NULL, "Not making");}
 //               io.c - START
 ///////////////////////////////////////////////
 
-static any read0(bool);
 
-static char Delim[] = " \t\n\r\"'(),[]`~{}";
-
-/* Buffer size */
-int bufSize(any x)
-{
-    return symBytes(x) + 1;
-}
-
-int pathSize(any x)
-{
-    int c = firstByte(x);
-
-    if (c != '@'  &&  (c != '+'))
-    {
-        return bufSize(x);
-    }
-    if (!Home)
-    {
-        return symBytes(x);
-    }
-    return strlen(Home) + symBytes(x);
-}
-
-void bufString(any x, char *p)
-{
-    int c, i;
-    uword w;
-
-    if (!isNil(x))
-    {
-        for (x = name(x), c = getByte1(&i, &w, &x); c; c = getByte(&i, &w, &x))
-        {
-            if (c == '^')
-            {
-                if ((c = getByte(&i, &w, &x)) == '?')
-                {
-                    c = 127;
-                }
-                else
-                {
-                    c &= 0x1F;
-                }
-            }
-            *p++ = c;
-        }
-    }
-    *p = '\0';
-}
-
-
-/*** Reading ***/
-void getStdin(void)
-{
-    Chr = getc(InFile);
-}
-
-
-/* Skip White Space and Comments */
-static int skipc(int c)
-{
-    if (Chr < 0)
-    {
-        return Chr;
-    }
-    for (;;)
-    {
-        while (Chr <= ' ')
-        {
-            Env.get();
-            if (Chr < 0)
-            {
-                return Chr;
-            }
-        }
-        if (Chr != c)
-        {
-            return Chr;
-        }
-        Env.get();
-        while (Chr != '\n')
-        {
-            if (Chr < 0)
-            {
-                return Chr;
-            }
-            Env.get();
-        }
-    }
-}
-
-static void comment(void)
-{
-    Env.get();
-    if (Chr != '{')
-    {
-        while (Chr != '\n')
-        {
-            if (Chr < 0)
-            {
-                return;
-            }
-            Env.get();
-        }
-    }
-    else
-    {
-        int n = 0;
-
-        for (;;) {  // #{block-comment}# from Kriangkrai Soatthiyanont
-            Env.get();
-            if (Chr < 0)
-            {
-                return;
-            }
-            if (Chr == '#'  &&  (Env.get(), Chr == '{'))
-            {
-                ++n;
-            }
-            else if (Chr == '}'  &&  (Env.get(), Chr == '#')  &&  --n < 0)
-            {
-                break;
-            }
-        }
-        Env.get();
-    }
-}
-
-static int skip(void)
-{
-    for (;;)
-    {
-        if (Chr < 0)
-        {
-            return Chr;
-        }
-        while (Chr <= ' ')
-        {
-            Env.get();
-            if (Chr < 0)
-            {
-                return Chr;
-            }
-        }
-
-        if (Chr != '#')
-        {
-            return Chr;
-        }
-        comment();
-    }
-}
-
-/* Test for escaped characters */
-static bool testEsc(void)
-{
-    for (;;)
-    {
-        if (Chr < 0)
-            return NO;
-        if (Chr != '\\')
-            return YES;
-        if (Env.get(), Chr != '\n')
-            return YES;
-        do
-        {
-            Env.get();
-        }
-        while (Chr == ' '  ||  Chr == '\t');
-    }
-}
-
-/* Read a list */
-static any rdList(void)
-{
-    any x;
-    cell c1;
-
-    for (;;)
-    {
-        if (skip() == ')')
-        {
-            Env.get();
-            return Nil;
-        }
-        if (Chr == ']')
-        {
-            return Nil;
-        }
-        if (Chr != '~')
-        {
-            x = cons(read0(NO),Nil);
-            Push(c1, x);
-            break;
-        }
-        Env.get();
-
-        x = read0(NO);
-        Push(c1, x);
-        if (isCell(x = data(c1) = EVAL(data(c1))))
-        {
-            while (isCell(cdr(x)))
-            {
-                x = cdr(x);
-            }
-            break;
-        }
-        drop(c1);
-    }
-
-    for (;;)
-    {
-        if (skip() == ')')
-        {
-            Env.get();
-            break;
-        }
-        if (Chr == ']')
-            break;
-        if (Chr == '.')
-        {
-            Env.get();
-            cdr(x) = skip()==')' || Chr==']'? data(c1) : read0(NO);
-            if (skip() == ')')
-                Env.get();
-            else if (Chr != ']')
-                err(NULL, x, "Bad dotted pair");
-            break;
-        }
-        if (Chr != '~')
-        {
-            x = cdr(x) = cons(read0(NO),Nil);
-        }
-        else
-        {
-            Env.get();
-            cdr(x) = read0(NO);
-            cdr(x) = EVAL(cdr(x));
-            while (isCell(cdr(x)))
-            {
-                x = cdr(x);
-            }
-        }
-    }
-    return Pop(c1);
-}
-
-/* Read one expression */
-static any read0(bool top)
-{
-    int i;
-    uword w;
-    any x, y;
-    cell c1, *p;
-
-    if (skip() < 0)
-    {
-        if (top)
-            return Nil;
-        eofErr();
-    }
-    if (Chr == '(')
-    {
-        Env.get();
-        x = rdList();
-        if (top  &&  Chr == ']')
-            Env.get();
-        return x;
-    }
-    if (Chr == '[')
-    {
-        Env.get();
-        x = rdList();
-        if (Chr != ']')
-            err(NULL, x, "Super parentheses mismatch");
-        Env.get();
-        return x;
-    }
-    if (Chr == '\'')
-    {
-        Env.get();
-        return cons(doQuote_D, read0(top));
-    }
-    if (Chr == ',')
-    {
-        Env.get();
-        return read0(top);
-    }
-    if (Chr == '`')
-    {
-        Env.get();
-        Push(c1, read0(top));
-        x = EVAL(data(c1));
-        drop(c1);
-        return x;
-    }
-    if (Chr == '"')
-    {
-        Env.get();
-        if (Chr == '"')
-        {
-            Env.get();
-            return Nil;
-        }
-        if (!testEsc())
-            eofErr();
-        putByte1(Chr, &i, &w, &p);
-        while (Env.get(), Chr != '"')
-        {
-            if (!testEsc())
-                eofErr();
-            putByte(Chr, &i, &w, &p, &c1);
-        }
-        y = popSym(i, w, p, &c1),  Env.get();
-        if (x = isIntern(tail(y), Transient))
-            return x;
-        intern(y, Transient);
-        return y;
-    }
-    if (strchr(Delim, Chr))
-        err(NULL, NULL, "Bad input '%c' (%d)", isprint(Chr)? Chr:'?', Chr);
-    if (Chr == '\\')
-        Env.get();
-    putByte1(Chr, &i, &w, &p);
-
-    int count=0;
-    for (;;)
-    {
-        count++;
-        // if (count > 6)
-        // {
-        //     printf("%s too long\n", (char*)&w);
-        //     bye(0);
-        // }
-        Env.get();
-        if (strchr(Delim, Chr))
-        {
-            break;
-        }
-        if (Chr == '\\')
-        {
-            Env.get();
-        }
-        putByte(Chr, &i, &w, &p, &c1);
-    }
-
-    y = popSym(i, w, p, &c1);
-    //printf("%p --> CAR = %p CDR = %p \n", y, y->car, y->cdr);
-    if (x = symToNum(tail(y), 0, '.', 0))
-    {
-        return x;
-    }
-    if (x = isIntern(tail(y), Intern))
-    {
-        return x;
-    }
-
-    intern(y, Intern);
-    //val(y) = Nil;
-    return y;
-}
-
-any read1(int end)
-{
-   if (!Chr)
-      Env.get();
-   if (Chr == end)
-      return Nil;
-   return read0(YES);
-}
-
-/* Read one token */
-any token(any x, int c)
-{
-    int i;
-    uword w;
-    any y;
-    cell c1, *p;
-
-    if (!Chr)
-        Env.get();
-    if (skipc(c) < 0)
-        return Nil;
-    if (Chr == '"')
-    {
-        Env.get();
-        if (Chr == '"')
-        {
-            Env.get();
-            return Nil;
-        }
-        if (!testEsc())
-            return Nil;
-        Push(c1, y =  cons(mkChar(Chr), Nil));
-        while (Env.get(), Chr != '"' && testEsc())
-            y = cdr(y) = cons(mkChar(Chr), Nil);
-        Env.get();
-        return Pop(c1);
-    }
-    if (Chr >= '0' && Chr <= '9')
-    {
-        putByte1(Chr, &i, &w, &p);
-        while (Env.get(), Chr >= '0' && Chr <= '9' || Chr == '.')
-            putByte(Chr, &i, &w, &p, &c1);
-        return symToNum(tail(popSym(i, w, p, &c1)), 0, '.', 0);
-    }
-    if (Chr != '+' && Chr != '-')
-    {
-        // TODO check what needs to be done about stack - FREE MUST BE ADDED
-        // char nm[bufSize(x)];
-        char *nm = (char *)malloc(bufSize(x));
-
-        bufString(x, nm);
-        if (Chr >= 'A' && Chr <= 'Z' || Chr == '\\' || Chr >= 'a' && Chr <= 'z' || strchr(nm,Chr))
-        {
-            if (Chr == '\\')
-                Env.get();
-            putByte1(Chr, &i, &w, &p);
-            while (Env.get(),
-                    Chr >= '0' && Chr <= '9' || Chr >= 'A' && Chr <= 'Z' ||
-                    Chr == '\\' || Chr >= 'a' && Chr <= 'z' || strchr(nm,Chr) )
-            {
-                if (Chr == '\\')
-                    Env.get();
-                putByte(Chr, &i, &w, &p, &c1);
-            }
-            y = popSym(i, w, p, &c1);
-            if (x = isIntern(tail(y), Intern))
-            {
-                free(nm);
-                return x;
-            }
-            intern(y, Intern);
-            val(y) = Nil;
-            free(nm);
-            return y;
-        }
-    }
-    y = mkTxt(c = Chr);
-    Env.get();
-    return mkChar(c);
-}
-
-bool eol(void)
-{
-   if (Chr < 0)
-      return YES;
-   if (Chr == '\n')
-   {
-      Chr = 0;
-      return YES;
-   }
-   if (Chr == '\r')
-   {
-      Env.get();
-      if (Chr == '\n')
-         Chr = 0;
-      return YES;
-   }
-   return NO;
-}
 
 any load(any ex, int pr, any x)
 {
