+++ src/lisp.h
@@ -335,6 +335,7 @@ any xSym(any);
 any doPrin(any x);
 void openErr(any ex, char *s);
 void eofErr(void);
+void comment(void);
 
 any doQuote(any x);
 any doEq(any x);
